% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMFData.R
\name{CompactDataMethod}
\alias{CompactDataMethod}
\title{Make query to the database}
\usage{
CompactDataMethod(databaseID, query = NULL, startdate = "2001-01-01",
  enddate = "2001-12-31", checkquery = FALSE)
}
\arguments{
\item{databaseID}{string. Database ID of the dataset. Obtained from \code{DataflowMethod}}

\item{startdate}{string. Start date in format of "YYYY-mm-dd".}

\item{enddate}{string. End date in format of "YYYY-mm-dd".}

\item{checkquery}{logical. If true, it will check the database ID is available or not.}

\item{querry}{list. A list that contains. If NULL, it will get the result of the first 3000 codes in the database.}
}
\value{
A list for .
The name of the list is the dimension name.
Each element of the list is a data frame with two columns; dimension code and dimension text(description).
}
\description{
\code{CompactDataMethod}
}
\examples{
IFS.available.codes <- DataStructureMethod('IFS')
databaseID <- 'IFS'
startdate='2001-01-01'
enddate='2016-12-31'
checkquery = FALSE
query <- list(CL_FREA="", CL_AREA_IFS="GR", CL_INDICATOR_IFS =c("NGDP_EUR","NGDP_XDC"))
NGDPquery <- CompactDataMethod(databaseID, query, startdate, enddate, checkquery)
NGDPquery[,1:5]
NGDPquery$Obs[[1]]

query <- list(CL_FREA="Q", CL_AREA_IFS="GR", CL_INDICATOR_IFS =c("NGDP_EUR","NGDP_XDC"))
NGDPquery <- CompactDataMethod(databaseID, query, startdate, enddate, checkquery)
NGDPquery[,1:5]
NGDPquery$Obs[[1]]

query <- list(CL_FREA="Q", CL_AREA_IFS="US", CL_INDICATOR_IFS = "")
NGDPquery <- CompactDataMethod(databaseID, query, startdate, enddate, checkquery)
NGDPquery[,1:5]
NGDPquery$Obs[[1]]
NGDPquery$`@INDICATOR`[grep("GDP", NGDPquery$`@INDICATOR`)]

# If the return data has only multiple data set, return a data frame
query <- list(CL_FREA="Q", CL_AREA_IFS="US", CL_INDICATOR_IFS =ALLGDPCodeValue[5:21]) # Looks like I can have 17 code at once maximum
NGDPquery <- CompactDataMethod(databaseID, query, startdate, enddate, checkquery)
NGDPquery[,1:5]
NGDPquery$Obs[[5]]

## Example US GDP
# If the return data has only one data set, return a list
query <- list(CL_FREA="Q", CL_AREA_IFS="US", CL_INDICATOR_IFS ="NGDP_SA_AR_XDC")
NGDPquery <- CompactDataMethod(databaseID, query, startdate, enddate, checkquery)
NGDPquery$Obs

DOT.available.codes <- DataStructureMethod('DOT')
names(DOT.available.codes)
DOT.available.codes[[2]]
CodeSearch(DOT.available.codes, "CL_AREA_DOT", "Taiwan")
query <- list(CL_FREQ = "", CL_AREA_DOT="F6", CL_INDICATOR_DOT = "", CL_COUNTERPART_AREA_DOT="")
DOTTXGquery <- CompactDataMethod('DOT', query, startdate, enddate, FALSE)
class(DOTTXGquery[1,])
DOTTXGquery$Obs[1]
}

