% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMFData.R
\name{CodeSearch}
\alias{CodeSearch}
\title{Code Search}
\usage{
CodeSearch(available.codes, code, searchtext, search.value = TRUE,
  search.text = TRUE)
}
\arguments{
\item{available.codes}{string. Database ID of the dataset.}

\item{code}{string. If true, it will check the database ID is available or not.}

\item{searchtext}{string. Database ID of the dataset.}

\item{search.value}{logical. If true, it will search only the .}

\item{search.text}{logical. If true, it will search .}
}
\value{
A list for each dimension.
The name of the list is the dimension name.
Each element of the list is a data frame with two columns; dimension code and dimension text(description).
}
\description{
\code{CodeSearch} search
}
\examples{
IFS.available.codes <- DataStructureMethod('IFS')
names(IFS.available.codes)
IFS.available.codes[[1]]
CodeSearch(IFS.available.codes, 'CLL', 'GDP')
CodeSearch(IFS.available.codes, 'CL_INDICATOR_IFS', 'GDP')
nrow(CodeSearch(IFS.available.codes, 'CL_INDICATOR_IFS', 'GDP'))
nrow(CodeSearch(IFS.available.codes, 'CL_INDICATOR_IFS', 'employ'))
nrow(CodeSearch(IFS.available.codes, 'CL_INDICATOR_IFS', 'employaaaaaaaaa'))
}

