% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMFData.R
\name{DataflowMethod}
\alias{DataflowMethod}
\title{Get List of Available Datasets from API}
\usage{
DataflowMethod()
}
\value{
\code{DataflowMethod} returns a data frame object to describe
available datasets from IMF data API. This data frame includes the following
two columns:

DatabaseID - Database ID uses for making API call \cr
DatabaseText - Database description
}
\description{
\code{DataflowMethod} returns a data frame with availble
datasets
}
\examples{
\donttest{
availableDB <- DataflowMethod()
availableDB
availableDB$DatabaseID[1]
}
}

