% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMFData.R
\name{DataflowMethod}
\alias{DataflowMethod}
\title{Get list of available datasets from API}
\usage{
DataflowMethod()
}
\value{
\code{DataflowMethod} returns a data frame object to describe
available datasets from the API. The data frame includes the following
components
DatabaseID Database ID uses for making API call
DatabaseText Database description.
}
\description{
\code{DataflowMethod} returns a data frame with availble
datasets
}
\examples{
availableDB <- DataflowMethod()
availableDB
availableDB$DatabaseID[1]

}

